+++ libimagentrytimetrack/src/lib.rs
@@ -33,7 +33,7 @@ extern crate libimagentrytag;
 
 mod constants;
 pub mod error;
-pub mod event;
+pub mod timetracking;
 pub mod eventstore;
 pub mod iter;
 pub mod result;
